+++ CHANGELOG.md
@@ -64,6 +64,7 @@
 * fix(Table): empty cell value causing hydration errors
 * fix(FileUpload): stuck focus while tabbing
 * fix(FileUpload): use native img element for blob URLs preview
+* fix(Modal/Slideover): remove close autofocus prevent
 * docs(app): implement AI search
 * doc(Component): added components for displaying props, slots, etc
 * chore(style): fix var(--b24ui-header-h`eight)